Improve design
Remove only files that will need to be saved again, add kind to file
name

+++ src/rx_classifier_scripts/train_classifier.py
@@ -128,20 +128,25 @@ def _get_args():
     Data.plot_only   = args.plot_only
     Data.load_trained= args.load_trained
 #---------------------------------
-def _merge_dataframes(l_rdf : list[RDataFrame]) -> RDataFrame:
+def _merge_dataframes(
+        l_rdf : list[RDataFrame],
+        kind  : str) -> RDataFrame:
     '''
     Takes list of dataframes, one for a different sample, after selection
+
+    l_rdf: List of ROOT dataframes to merge
+    kind : Type of sample, e.g. sig, bkg
     '''
 
-    for fpath in glob.glob(f'{Data.cache_dir}/*.root'):
-        shutil.rmtree(fpath)
-
     for i_rdf, rdf, in enumerate(l_rdf):
-        fpath = f'{Data.cache_dir}/file_{i_rdf:03}.root'
+        fpath = f'{Data.cache_dir}/{kind}_{i_rdf:03}.root'
+        if os.path.isfile(fpath):
+            shutil.rmtree(fpath)
+
         log.info(f'Saving temporary file to: {fpath}')
         rdf.Snapshot('tree', fpath)
 
-    rdf = RDataFrame('tree', f'{Data.cache_dir}/file_*.root')
+    rdf = RDataFrame('tree', f'{Data.cache_dir}/{kind}_*.root')
 
     return rdf
 #---------------------------------
